V2/architecture/pyproject.toml
[tool.poetry]
name = "Ergodeon-agent-system"
version = "1.0.0"
description = "Ergodeon Agent System - AI-powered software development agents"
authors = ["Ergodeon Team"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/wku/Ergodeon.git"
repository = "https://github.com/wku/Ergodeon.git"
keywords = ["ai", "agents", "development", "automation", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[tool.poetry.dependencies]
python = "^3.11"
asyncio = "^3.4.3"
aiofiles = "^23.2.1"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
chromadb = "^0.4.22"
litellm = "^1.17.0"
openai = "^1.10.0"
anthropic = "^0.18.0"
pyyaml = "^6.0.1"

# Optional dependencies
fastapi = {version = "^0.109.0", optional = true}
uvicorn = {version = "^0.27.0", extras = ["standard"], optional = true}
redis = {version = "^5.0.1", optional = true}
streamlit = {version = "^1.30.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
black = "^24.1.0"
ruff = "^0.1.14"
mypy = "^1.8.0"
isort = "^5.13.2"
pre-commit = "^3.6.0"

[tool.poetry.extras]
api = ["fastapi", "uvicorn"]
redis = ["redis"]
ui = ["streamlit"]
all = ["fastapi", "uvicorn", "redis", "streamlit"]

[tool.poetry.scripts]
Ergodeon = "src.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers --cov=. --cov-report=html --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["."]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/env/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
